add logs, change migr

+++ internal/db/migrations/sql/00001_init_db.sql
@@ -1,12 +1,12 @@
 -- +goose Up
 CREATE TABLE Users (
-    userID bigint primary key generated always as identity,
+    userID int primary key generated always as identity,
     userLogin varchar(200) not null unique,
     userPassword varchar(200) not null
 );
 
 CREATE TABLE Orders (
-    userID bigint not null,
+    userID int not null,
     orderNumber varchar(200) not null unique,
     orderStatus varchar(20),
     accrual real default 0,
@@ -14,14 +14,14 @@ CREATE TABLE Orders (
 );
 
 CREATE TABLE OrdersOperations (
-    userID bigint not null,
+    userID int not null,
     orderNumber varchar(200) not null,
     pointsQuantity real default 0,
     processedAt timestamp default NULL
 );
 
 CREATE TABLE UsersBalance (
-    userID bigint not null,
+    userID int not null,
     pointsSum real default 0,
     pointsLoss real default 0
 );
